split and change snapshot tests to medium
